requirements.txt
Flask
mysql-connector-python
gunicorn
qrcode
opencv-python-headless
numpy
face-recognition
Pillow